try to sort the arch_init_freemem
