+++ sel4_vspace/src/arch/mod.rs
@@ -1,6 +1,9 @@
 #[cfg(target_arch = "aarch64")]
 mod aarch64;
+#[cfg(target_arch = "aarch64")]
+pub use aarch64::*;
+
 #[cfg(target_arch = "riscv64")]
 mod riscv64;
-
+#[cfg(target_arch = "riscv64")]
 pub use riscv64::*;
